+++ resources/views/container/posts/filters/created.blade.php
@@ -5,14 +5,14 @@
         <div class="col-md-6">
             <div class="form-group">
                 <input type='text' name="start_created_at" id='start_created_at'
-                       value="{{$request->get('start_created_at')}}" class="form-control"/>
+                       value="{{$request->get('start_created_at')}}" class="form-control datetimepicker"/>
             </div>
         </div>
 
         <div class="col-md-6">
             <div class="form-group">
                 <input type='text' name="end_created_at" id='end_created_at' value="{{$request->get('end_created_at')}}"
-                       class="form-control"/>
+                       class="form-control datetimepicker"/>
             </div>
         </div>
     </div>