+++ resources/views/container/tools/category/info.blade.php
@@ -1,8 +1,7 @@
 <div class="wrapper-md">
 
     <input type="hidden" name="taxonomy" value="category">
-    <input type="hidden" name="term_id" @if($termTaxonomy->exists)value="{{$termTaxonomy->term->id}}"
-           @else value="0" @endif>
+    <input type="hidden" name="term_id" @if($termTaxonomy->exists)value="{{$termTaxonomy->term->id}}" @else value="0" @endif>
 
     <div class="form-group">
         <label class="col-sm-2 control-label">{{trans('cms::tools/category.parent')}}</label>
@@ -17,7 +16,8 @@
 
                     <option value="{{$value->id}}"
                             @if($termTaxonomy->exists && $termTaxonomy->parent_id == $value->id) selected @endif >
-                        {{$value->term->getContent('name')}}</option>
+                            {{$value->term->getContent('name')}}
+                    </option>
                 @endforeach
             </select>
         </div>
@@ -27,8 +27,9 @@
     <div class="form-group m-t-md">
         <label class="col-sm-2 control-label">{{trans('cms::tools/category.slug')}}</label>
         <div class="col-sm-10">
-            <input type="text" name="slug" @if($termTaxonomy->exists) value="{{$termTaxonomy->term->slug}}"
-                   @endif required class="form-control" placeholder="news">
+            <input type="text" name="slug"
+                   @if($termTaxonomy->exists) value="{{$termTaxonomy->term->slug}}" @endif
+                   required class="form-control" placeholder="news">
         </div>
     </div>