+++ resources/views/fields/code.blade.php
@@ -1,9 +1,9 @@
 <div class="form-group{{ $errors->has($name) ? ' has-error' : '' }}">
     @if(isset($title))
-        <label for="field-{{$slug}}">{{$title}}</label>
+        <label for="field-{{$lang}}-{{$slug}}">{{$title}}</label>
     @endif
-    <div id="ace-code-block-{{$slug}}" style="width: 100%; min-height: 300px;"></div>
-    <input type="hidden" class="form-control {{$class or ''}}" id="field-{{$slug}}"
+    <div id="ace-code-block-{{$lang}}-{{$slug}}" style="width: 100%; min-height: 300px;"></div>
+    <input type="hidden" class="form-control {{$class or ''}}" id="field-{{$lang}}-{{$slug}}"
            @if(isset($prefix))
            name="{{$prefix}}[{{$lang}}]{{$name}}"
            @else
@@ -19,16 +19,16 @@
 @push('scripts')
 <script>
     $(function () {
-        var editor{{$slug}} = ace.edit('ace-code-block-{{$slug}}');
-        editor{{$slug}}.getSession().setMode('ace/mode/javascript');
-        editor{{$slug}}.setTheme('ace/theme/monokai');
+        var editor{{$lang}}{{$slug}} = ace.edit('ace-code-block-{{$lang}}-{{$slug}}');
+        editor{{$lang}}{{$slug}}.getSession().setMode('ace/mode/javascript');
+        editor{{$lang}}{{$slug}}.setTheme('ace/theme/monokai');
+        editor{{$lang}}{{$slug}}.getSession().setUseWorker(false);
 
-        var input{{$slug}} = $('#field-{{$slug}}');
-        editor{{$slug}}.getSession().setValue(input{{$slug}}.val());
-        editor{{$slug}}.getSession().on('change', function () {
-            input{{$slug}}.val(editor{{$slug}}.getSession().getValue());
+        var input{{$lang}}{{$slug}} = $('#field-{{$lang}}-{{$slug}}');
+        editor{{$lang}}{{$slug}}.getSession().setValue(input{{$lang}}{{$slug}}.val());
+        editor{{$lang}}{{$slug}}.getSession().on('change', function () {
+            input{{$lang}}{{$slug}}.val(editor{{$lang}}{{$slug}}.getSession().getValue());
         });
-
     });
 </script>
 @endpush