+++ resources/views/fields/list.blade.php
@@ -1,6 +1,6 @@
 <div class="form-group input-sort">
     <label>{{$title}}</label>
-   <ul id="sortable-{{$slug}}" class="container-fluid">
+   <ul id="sortable-{{$slug}}" class="container-fluid dd-list">
 
        @if(!isset($value) || is_null($value))
                <li class="ui-state-default form-group row">
@@ -31,14 +31,11 @@
        @endif
     </ul>
     <div class="button-group text-center">
-        <button onclick="newitem{{$slug}}()" class="btn btn-xs alert-info">Add new</button>
+        <button onclick="newitem{{$slug}}()" class="btn btn-xs alert-info">new</button>
     </div>
 </div>
 <div class="line line-dashed b-b line-lg"></div>
 <style>
-    .input-sort li{
-        list-style: none;
-    }
     .input-sort .form-control {
         width: 83.33333333%;
     }