+++ changelog.md
@@ -1,4 +1,10 @@
-# 2.0 (dev) (17.09.2017)
+# 1.2 (dev) (17.09.2017-25.09.2017)
 
 - Added full text search
 - Added fields
+- Added set locale behavior
+- Added attachment any model
+- Added headless mode
+- Added min/minlength, max/maxlength to input&textarea
+
+- Fix code editor language